move test from expiration to core test
